+++ internal/repositories/versions/get_versions.go
@@ -14,11 +14,11 @@ func (r *Repository) GetVersions(ctx context.Context, filter models.GetVersionsF
 		From("versions")
 
 	if filter.Limit > 0 {
-		query = query.Limit(uint64(filter.Limit))
+		query = query.Limit(filter.Limit)
 	}
 
 	if filter.Offset > 0 {
-		query = query.Offset(uint64(filter.Offset))
+		query = query.Offset(filter.Offset)
 	}
 
 	if filter.ProjectID > 0 {
